Styling improvements on List View
